Changed the names and paths of the testcases and helper files
